Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/execution-timing/030-expected.txt

FAILED (This TC requires JavaScript enabled)

FAIL  scheduler: javascript: URL in HREF, onclick handler assert_array_equals: expected property 2 to be "end script #1" but got "JS URL" (expected array ["inline script #1", "click event", "end script #1", "JS URL", "inline script #2"] got ["inline script #1", "click event", "JS URL", "end script #1", "inline script #2"])

assert_array_equals: expected property 2 to be "end script #1" but got "JS URL" (expected array ["inline script #1", "click event", "end script #1", "inline script #2", "JS URL"] got ["inline script #1", "click event", "JS URL", "end script #1", "inline script #2"])
